audio: correct dependencies for compilation with AOSP flag
Add liblog to the dependency list for voice_processing and
disable compilation of audiod when AOSP flag is set.

+++ hal/msm8974/platform.c
@@ -3221,10 +3221,10 @@ snd_device_t platform_get_input_snd_device(void *platform, audio_devices_t out_d
                     (my_data->source_mic_type & SOURCE_DUAL_MIC)) {
                     snd_device = SND_DEVICE_IN_UNPROCESSED_STEREO_MIC;
                 }
-                else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_3) &&
+                else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_3) &&
                          (my_data->source_mic_type & SOURCE_THREE_MIC)) {
                          snd_device = SND_DEVICE_IN_UNPROCESSED_THREE_MIC;
-               } else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_4) &&
+               } else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_4) &&
                           (my_data->source_mic_type & SOURCE_QUAD_MIC)) {
                           snd_device = SND_DEVICE_IN_UNPROCESSED_QUAD_MIC;
                } else {
@@ -3250,10 +3250,10 @@ snd_device_t platform_get_input_snd_device(void *platform, audio_devices_t out_d
                        (channel_mask == AUDIO_CHANNEL_IN_STEREO)) &&
                        (my_data->source_mic_type & SOURCE_DUAL_MIC)) {
                 snd_device = SND_DEVICE_IN_VOICE_REC_DMIC_STEREO;
-            } else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_3) &&
+            } else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_3) &&
                        (my_data->source_mic_type & SOURCE_THREE_MIC)) {
                 snd_device = SND_DEVICE_IN_THREE_MIC;
-            } else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_4) &&
+            } else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_4) &&
                        (my_data->source_mic_type & SOURCE_QUAD_MIC)) {
                 snd_device = SND_DEVICE_IN_QUAD_MIC;
             }
@@ -3270,10 +3270,10 @@ snd_device_t platform_get_input_snd_device(void *platform, audio_devices_t out_d
                  (channel_mask == AUDIO_CHANNEL_IN_STEREO)) &&
                  (my_data->source_mic_type & SOURCE_DUAL_MIC)) {
                  snd_device = SND_DEVICE_IN_UNPROCESSED_STEREO_MIC;
-             } else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_3) &&
+             } else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_3) &&
                  (my_data->source_mic_type & SOURCE_THREE_MIC)) {
                  snd_device = SND_DEVICE_IN_UNPROCESSED_THREE_MIC;
-             } else if (((int)channel_mask == AUDIO_CHANNEL_INDEX_MASK_4) &&
+             } else if (((int)channel_mask == (int)AUDIO_CHANNEL_INDEX_MASK_4) &&
                  (my_data->source_mic_type & SOURCE_QUAD_MIC)) {
                  snd_device = SND_DEVICE_IN_UNPROCESSED_QUAD_MIC;
              } else {
